Refactor chat history to endpoint
